Lower-level access to the RELAX-IV algorithm

`.RELAX_IV` provides low-level access to the FORTRAN; taking the exact input as the FORTRAN code. It does NOT do any input checking.

.RELAX_IV(n1, na1, startn1, endn1, c1, u1, b1, rc1, crash1, large1)

Arguments

n1

Number of nodes

na1

Number of arcs

startn1

Starting nodes

endn1

Ending nodes

c1

Costs vector

u1

Capacities

b1

Demands

rc1

Reduced Costs

crash1

Set crash equal to 1 to activate an auction/shortest path subroutine for getting the initial price-flow pair. this is recommended for difficult problems where the default initialization yields long solution times.

large1

A very large integer to represent infinity

Value

A list containing all parameters, as well as `x1` which contains the solution, and `feasible1` which contains a flag for feasibility.