Display license information about embedded code

Function to display license information regarding code embedded in rrelaxiv.

Usage

relaxinfo()

Value

Some information about licenses of code and algorithms.

Examples

relaxinfo()
#> The rrelaxiv package implements D. P. Bertsekas
#>  and P. Tseng's RELAX-IV algorithm and code:
#> 
#>  Bersekas, D. P. and  Tseng, P., "Relaxation Methods for
#>  Minimum Cost ..." Operations Research, vol. 26, 1988, 93-114;
#>  Bertsekas, D. P., "An Auction/Sequential Shortest Path
#>  Algorithm for the Minimum Cost Flow Problem", LIDS Report
#>  P-2146, MIT, Nov. 1992;
#>  Bertsekas, D.P., Linear Network Optimization, MIT Press, 1991;
#>  Bertsekas, D.P., Network Optimization: Continuous and Discrete
#>  Models, Athena Scientific, 1998;
#>  <http://web.mit.edu/dimitrib/www/noc.htm>.
#> 
#>  Bertsekas and Tseng freely permit their software to be used for
#>  research purposes, but non-research uses, including the use of it
#>  to 'satisfy in any part commercial delivery requirements to
#>  government or industry,' require a special agreement with them.
#>  By extension, this requirement applies to most any use of R
#>  functions in the rrelaxiv package.
#> 
#>  To request permissions not here relayed, contact Professor 
#>  Bertsekas at Laboratory for Information and Decision Systems
#>  Massachusetts Institute of Technology
#>  Cambridge, MA 02139
#>  (617) 253-7267 <dimitrib@mit.edu>